Java Utility Class [3.30.2023]
This is just a utility class that I wrote today because constantly using For Loops to display both normal Arrays and 2D. The absolute value method isn't that helpful, but I thought, "Why not?". Same with the print() and println. It doesn't shorten System.out.print() that much, but util.print() is still a little shorter. Also, comments are... lacking 😆, sorry.